feat(tipo-ato): peso de complexidade decimal e tempo de referência (RF-34a, RF-34f, RF-46c)
Fatia 5 do Dashboard v2 (PLANO-dashboard-v2.md), com as decisões do dono de 25/09/2026.

Peso de complexidade vira decimal 0,50–2,50 em passos de 0,05 (PesoDeComplexidade, Domain).
A migration ConverteTempoDeReferenciaEPesoDecimalEmTiposAto CONVERTE o dado existente no USING
do ALTER COLUMN (1→1,00, 2→1,25, 3→1,50, 4→1,75, 5→2,00; ≤ 0 → 0,50; ≥ 6 → 2,50) — o
AlterColumn gerado pelo EF faria só o cast do Postgres (3 → 3,00, fora da faixa). Os CHECK
entram depois da conversão; o Down volta pelo mapa inverso. `decimal` (não `double`) porque é
base 10: 1,35 % 0,05 é exatamente 0.

Tempo de referência (ADR-0043): só o valor informado é gravado (tempo_referencia_minutos,
2–240, nulo = não informado). A referência efetiva — informado → mediana dos últimos 12 meses
com ≥ 30 conferências válidas (descarte > 4× a estimativa) → estimativa
round(TempoMedioPorAtoMinutos × peso) — é calculada na leitura por TempoDeReferencia (Domain,
arredondamento escolar: Math.Round é bancário por padrão, 22,5 daria 22). Sem cache: a mediana
muda a cada conclusão/correção/reabertura/ajuste e não há job pra recalcular.

Protocolo.Duracao não é coluna, então o histórico vem de uma query projetada por listagem
(ObterDuracoesConcluidasPorTipoAsync: tipo, início/fim, ciclos, último ajuste), somada pela
mesma função da propriedade (Protocolo.CalcularDuracao) — sem N+1 e sem reescrever a conta em
SQL. GET /tipos-ato/com-uso calcula só pros tipos da página.

Contrato: GET /tipos-ato ganha pesoComplexidade (RF-18a); GET /tipos-ato/com-uso ganha
pesoComplexidade decimal e tempoReferencia { minutos, origem, informadoMinutos, medianaMinutos,
conferenciasNoHistorico }; PUT /tipos-ato/{id}/peso aceita decimal e devolve 400 { motivo }
fora da faixa/passo (sai o clamp silencioso); novo PUT /tipos-ato/{id}/tempo-referencia
{ minutos: int | null } só pro Administrador. complexidadeMedia do Dashboard passa a ser a
média dos pesos decimais.

+++ src/Dispatch.Application/CasosDeUso/ObterDashboard.cs
@@ -262,7 +262,7 @@ public sealed class ObterDashboard(
     {
         var pesos = protocolosDoConferente
             .Where(p => p.TipoAtoId is not null && catalogo.ContainsKey(p.TipoAtoId.Value))
-            .Select(p => catalogo[p.TipoAtoId!.Value].PesoComplexidade)
+            .Select(p => (double)catalogo[p.TipoAtoId!.Value].PesoComplexidade)
             .ToList();
         return pesos.Count == 0 ? 0 : pesos.Average();
     }
